+++ variables.tf
@@ -1,10 +1,20 @@
-variable "as_pre_prefix" {
+variable "pet_first" {
   default = false
   type    = bool
 
   description = <<-END
-    Add Segment as a Pre Prefix
-    https://go.s3d.club/tf/exrn/name-tags#as_pre_prefix
+    an option to include the pet name as the first component of the outputted name.
+    https://go.s3d.club/tf/name#pet_first
+    END
+}
+
+variable "context" {
+  default = null
+  type    = string
+
+  description = <<-END
+    a contextual component for the final prefix _(something like "ec2" works well here)_.
+    https://go.s3d.club/tf/name#context
     END
 }
 
@@ -13,28 +23,29 @@ variable "disable_date" {
   type    = bool
 
   description = <<-END
-    Disable Date Segment
-    https://go.s3d.club/tf/exrn/name-tags#disabe_date
+    an option to disable the date component.
+    https://go.s3d.club/tf/name#disable_date
     END
 }
 
-variable "name_prefix" {
-  default = null
-  type    = string
+variable "keepers" {
+  default = {}
+  type    = map(string)
 
   description = <<-END
-    Name Prefix
-    https://go.s3d.club/tf/exrn/name-tags#name_prefix
+    an map of extra keepers for the pet and thus the name.
+    https://go.s3d.club/tf/name#disable_date
     END
 }
 
-variable "name_segment" {
-  default = null
-  type    = string
+
+variable "length" {
+  default = 63 # AWS S3 Bucket names are limited to a max len of 63.
+  type    = number
 
   description = <<-END
-    Name Segment
-    https://go.s3d.club/tf/exrn/name-tags#name_segment
+    the maximum length of the name that will be returned.
+    https://go.s3d.club/tf/name#length
     END
 }
 
@@ -42,8 +53,8 @@ variable "path" {
   type = string
 
   description = <<-END
-    Path
-    https://go.s3d.club/tf/exrn/name-tags#path
+    the path for the module using the name.
+    https://go.s3d.club/tf/name#path
     END
 }
 
@@ -52,18 +63,8 @@ variable "pet_length" {
   type    = number
 
   description = <<-END
-    Pet Names
-    https://go.s3d.club/tf/exrn/name-tags#pet_length
-    END
-}
-
-variable "size" {
-  default = 64
-  type    = number
-
-  description = <<-END
-    Tags
-    https://go.s3d.club/tf/exrn/name-tags#size
+    the number of "pets" to include in the name
+    https://go.s3d.club/tf/name#pet_length
     END
 }
 
@@ -71,7 +72,7 @@ variable "tags" {
   type = map(string)
 
   description = <<-END
-    Tags
-    https://go.s3d.club/tf/exrn/name-tags#tags
+    the tags that will be included in output.
+    https://go.s3d.club/tf/name#tags
     END
 }