+++ CHANGES.md
@@ -4,5 +4,8 @@ Recently completed and planned work is tracked here.
 ## [0.0.1](.) [0.1.0](.)
 - Created early versions
 
-## [0.1.0](.)
-- Fixed `s3d-post-flow` for `v` tags
+## [0.1.1](.)
+- Added variable for `keepers`
+- Improved variables and outputs
+- Removed `s3d-post-flow`
+- Upgraded `change` module